move clustering functions.
